+++ benker/parsers/ooxml.py
@@ -15,6 +15,7 @@ import functools
 
 from lxml import etree
 
+from benker.cell import get_content_text
 from benker.parsers.base_parser import BaseParser
 from benker.parsers.base_parser import value_of as base_value_of
 from benker.parsers.lxml_iterwalk import iterwalk
@@ -1025,6 +1026,9 @@ class OoxmlParser(BaseParser):
 
         :type  w_tc: etree._Element
         :param w_tc: Table element.
+
+        .. versionchanged:: 0.4.4
+           Improved empty cells detection for Formex4 conversion (``<IE/>`` tag management).
         """
         state = self._state
 
@@ -1114,4 +1118,16 @@ class OoxmlParser(BaseParser):
             # todo: calculate the ``@rotate`` attribute.
 
             content = w_tc.xpath('w:p | w:tbl', namespaces=NS)
+
+            # The detection of empty cells (without text or image) is used when converting
+            # to the Formex4 format in order to insert an empty tag ``<IE/>``.
+            # see: https://github.com/laurent-laporte-pro/benker/issues/13
+            if (
+                not get_content_text(content)
+                and not w_tc.xpath("count(.//w:drawing)", namespaces=NS)
+                and not w_tc.xpath("count(.//w:pict)", namespaces=NS)
+            ):
+                # The cell has no text or image.
+                styles["x-cell-empty"] = "true"
+
             state.row.insert_cell(content, width=width, height=height, styles=styles)