+++ tests/parsers/test_ooxml_parser.py
@@ -7,17 +7,16 @@ from benker.parsers.ooxml import OoxmlParser
 from benker.table import Table
 
 TEST_DATA = [
-
+    # -------------
     # without style
     # -------------
-
     pytest.param(
         u"""<w:tc xmlns:w="http://schemas.openxmlformats.org/wordprocessingml/2006/main">
             <w:p w:rsidR="00EF2ECA" w:rsidRDefault="00EF2ECA"><w:r><w:t>empty</w:t></w:r></w:p>
         </w:tc>""",
         {},
-        id="no_style"),
-
+        id="no_style",
+    ),
     pytest.param(
         u"""<w:tc xmlns:w="http://schemas.openxmlformats.org/wordprocessingml/2006/main">
             <w:tcPr>
@@ -27,11 +26,11 @@ TEST_DATA = [
             <w:p w:rsidR="00EF2ECA" w:rsidRDefault="00EF2ECA"><w:r><w:t>B</w:t></w:r></w:p>
         </w:tc>""",
         {},
-        id="style_is_empty"),
-
+        id="style_is_empty",
+    ),
+    # -------
     # Borders
     # -------
-
     pytest.param(
         u"""<w:tc xmlns:w="http://schemas.openxmlformats.org/wordprocessingml/2006/main">
             <w:tcPr>
@@ -45,11 +44,9 @@ TEST_DATA = [
             </w:tcPr>
             <w:p w:rsidR="00EF2ECA" w:rsidRDefault="00EF2ECA"><w:r><w:t>1</w:t></w:r></w:p>
         </w:tc>""",
-        {'border-collapse': 'collapse',
-         'border-right': 'solid 1.5pt #4472C4',
-         'border-bottom': 'none'},
-        id="style_border-right_solid_border-bottom_none"),
-
+        {"border-collapse": "collapse", "border-right": "solid 1.5pt #4472C4", "border-bottom": "none"},
+        id="style_border-right_solid_border-bottom_none",
+    ),
     pytest.param(
         u"""<w:tc xmlns:w="http://schemas.openxmlformats.org/wordprocessingml/2006/main">
             <w:tcPr>
@@ -63,12 +60,14 @@ TEST_DATA = [
             </w:tcPr>
             <w:p w:rsidR="00EF2ECA" w:rsidRDefault="00EF2ECA"/>
         </w:tc>""",
-        {'border-top': 'double 1.0pt #FFFFFF',
-         'border-right': 'dotted 2.0pt #FF0000',
-         'border-bottom': 'dashed 3.0pt #00FF00',
-         'border-left': 'outset 4.0pt #0000FF'},
-        id="style_border-all"),
-
+        {
+            "border-top": "double 1.0pt #FFFFFF",
+            "border-right": "dotted 2.0pt #FF0000",
+            "border-bottom": "dashed 3.0pt #00FF00",
+            "border-left": "outset 4.0pt #0000FF",
+        },
+        id="style_border-all",
+    ),
     pytest.param(
         u"""<w:tc xmlns:w="http://schemas.openxmlformats.org/wordprocessingml/2006/main">
             <w:tcPr>
@@ -80,14 +79,16 @@ TEST_DATA = [
             </w:tcPr>
             <w:p w:rsidR="00EF2ECA" w:rsidRDefault="00EF2ECA"/>
         </w:tc>""",
-        {'border-collapse': 'collapse',
-         'x-border-tl2br': 'solid 0.5pt',
-         'x-border-tr2bl': 'solid 0.5pt'},
-        id="style_border-tl2br-tr2bl"),
-
+        {
+            "border-collapse": "collapse",
+            "x-border-tl2br": "solid 0.5pt",
+            "x-border-tr2bl": "solid 0.5pt",
+        },
+        id="style_border-tl2br-tr2bl",
+    ),
+    # ------------------------
     # align and vertical-align
     # ------------------------
-
     pytest.param(
         u"""<w:tc xmlns:w="http://schemas.openxmlformats.org/wordprocessingml/2006/main">
             <w:tcPr><w:vAlign w:val="top"/></w:tcPr>
@@ -96,9 +97,9 @@ TEST_DATA = [
                 <w:r><w:t>text</w:t></w:r>
             </w:p>
         </w:tc>""",
-        {'vertical-align': 'top', 'align': 'left'},
-        id="style_vertical-align_top_align_left"),
-
+        {"vertical-align": "top", "align": "left"},
+        id="style_vertical-align_top_align_left",
+    ),
     pytest.param(
         u"""<w:tc xmlns:w="http://schemas.openxmlformats.org/wordprocessingml/2006/main">
             <w:tcPr><w:vAlign w:val="bottom"/></w:tcPr>
@@ -107,9 +108,9 @@ TEST_DATA = [
                 <w:r><w:t>text</w:t></w:r>
             </w:p>
         </w:tc>""",
-        {'vertical-align': 'bottom', 'align': 'right'},
-        id="style_vertical-align_bottom_align_right"),
-
+        {"vertical-align": "bottom", "align": "right"},
+        id="style_vertical-align_bottom_align_right",
+    ),
     pytest.param(
         u"""<w:tc xmlns:w="http://schemas.openxmlformats.org/wordprocessingml/2006/main">
             <w:tcPr><w:vAlign w:val="center"/></w:tcPr>
@@ -118,9 +119,9 @@ TEST_DATA = [
                 <w:r><w:t>text</w:t></w:r>
             </w:p>
         </w:tc>""",
-        {'vertical-align': 'middle', 'align': 'center'},
-        id="style_vertical-align_middle_align_center"),
-
+        {"vertical-align": "middle", "align": "center"},
+        id="style_vertical-align_middle_align_center",
+    ),
     pytest.param(
         u"""<w:tc xmlns:w="http://schemas.openxmlformats.org/wordprocessingml/2006/main">
             <w:tcPr><w:vAlign w:val="both"/></w:tcPr>
@@ -129,13 +130,13 @@ TEST_DATA = [
                 <w:r><w:t>text</w:t></w:r>
             </w:p>
         </w:tc>""",
-        {'vertical-align': 'w-both', 'align': 'justify'},
-        id="style_vertical-align_both_align_justify"),
-
+        {"vertical-align": "w-both", "align": "justify"},
+        id="style_vertical-align_both_align_justify",
+    ),
 ]
 
 
-@pytest.mark.parametrize('w_tc_content, expected', TEST_DATA)
+@pytest.mark.parametrize("w_tc_content, expected", TEST_DATA)
 def test_parse_tc(w_tc_content, expected):
     builder = BaseBuilder()
     parser = OoxmlParser(builder)
@@ -153,4 +154,7 @@ def test_parse_tc(w_tc_content, expected):
     # -- check the styles
     table = state.table
     cell = table[(1, 1)]
-    assert expected == cell.styles
+
+    # Ignore cell styles extensions (like 'x-cell-empty').
+    actual = {k: v for k, v in cell.styles.items() if not k.startswith("x-cell-")}
+    assert expected == actual